[XPath] Improved internal documentation
